version 0.4.97

+++ src/calibre/ebooks/lrf/txt/convert_from.py
@@ -31,7 +31,8 @@ def fix_image_includes(tdir, match):
     p = os.path.dirname(dest)
     if not os.path.exists(p):
         os.makedirs(p)
-    shutil.copyfile(src, dest)
+    if not os.path.exists(dest):
+        shutil.copyfile(src, dest)
     
 
 def generate_html(txtfile, encoding, tdir):
